+++ src/main/kotlin/me/nobaboy/nobaaddons/core/fishing/TrophyFish.kt
@@ -1,13 +1,13 @@
 package me.nobaboy.nobaaddons.core.fishing
 
-import me.nobaboy.nobaaddons.core.ItemRarity
+import me.nobaboy.nobaaddons.core.Rarity
+import me.nobaboy.nobaaddons.core.Rarity.Companion.rarityFormatted
 import me.nobaboy.nobaaddons.repo.RepoObjectArray.Companion.listFromRepository
-import me.nobaboy.nobaaddons.utils.TextUtils.formatted
+import me.nobaboy.nobaaddons.utils.TextUtils.toText
 import net.minecraft.text.Text
-import net.minecraft.util.Formatting
 
-data class TrophyFish(val id: String, val name: String, val rarity: ItemRarity) {
-	val displayName: Text get() = name.formatted(rarity.formatting ?: Formatting.RESET)
+data class TrophyFish(val id: String, val name: String, val rarity: Rarity) {
+	val displayName: Text by lazy { name.toText().rarityFormatted(rarity) }
 
 	companion object {
 		val FISH by TrophyFish::class.listFromRepository("fishing/trophy_fish.json")