+++ src/main/kotlin/me/nobaboy/nobaaddons/features/chat/filters/mobs/SeaCreatureSpawnMessageChatFilter.kt
@@ -6,9 +6,6 @@ import me.nobaboy.nobaaddons.features.chat.filters.IChatFilter
 object SeaCreatureSpawnMessageChatFilter : IChatFilter {
 	override val enabled: Boolean get() = config.hideSeaCreatureSpawnMessage
 
-	override fun shouldFilter(message: String): Boolean {
-		val seaCreature = SeaCreature.getBySpawnMessage(message) ?: return false
-
-		return seaCreature.rarity.isAtMost(config.seaCreatureMaximumRarity)
-	}
+	override fun shouldFilter(message: String): Boolean =
+		SeaCreature.getBySpawnMessage(message)?.let { it.rarity <= config.seaCreatureMaximumRarity } == true
 }